refactor: extract sunlight as a custom DirectionalLight class
